Time: 151 ms (38.07%), Space: 87.1 MB (24.75%) - LeetHub

+++ 2064-minimized-maximum-of-products-distributed-to-any-store/2064-minimized-maximum-of-products-distributed-to-any-store.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    long long int findNoOfStores(vector<int>&quantities, int products){
-        long long int noOfStores=0;
+     int findNoOfStores(vector<int>&quantities, int products){
+         int noOfStores=0;
         for(int i=0; i<quantities.size();i++){
             noOfStores+=ceil((double)quantities[i]/(double)products);
         }
@@ -11,13 +11,13 @@ public:
     }
     int minimizedMaximum(int n, vector<int>& quantities) {
         // long long int total=accumulate(quantities.begin(), quantities.end(), 0);
-        long long int low=1;
-        long long int high=*max_element(quantities.begin(), quantities.end());
-        long long int ans=-1;
+         int low=1;
+         int high=*max_element(quantities.begin(), quantities.end());
+         int ans=-1;
         
         while(low<=high){
-            long long int mid=(low+high)/2;
-            long long int stores=findNoOfStores(quantities, mid);
+             int mid=(low+high)/2;
+             int stores=findNoOfStores(quantities, mid);
             
             if(stores > n){
                 low=mid+1;
